[Refactor] DRTimeLineView 컴포넌트 구현 및 적용 (#288)
